Worked on reservations and added 2 files
Added the ability to make reservations, however
it is not complete bc at the moment you can
only reserve one seat at a time. This differs
from the requirements of the project. Also added
the sample file which serves as a backup for the
seats file. The seats file for this commit is
in a blank state.